Handle cleanup and use
Decided against having a typedef for every handle type, mostly to make
it more clear that it isn't doing anything special and freeing the names
up for actual complicated classes (e.g. a Swapchain class that manages
everything it needs to itself). Also moved the current implementation
over to using the handles.