Vulkan struct initialisers
Now with a crap-load of helper functions that fully default initialise
all of the Vulkan structs, making sure everything has the right
VkStructureType and starting everything else at 0. This sort of thing is
in the Khronos group's examples so it's a bit odd it isn't a standard
thing?... Either way it seems to be done, but only tested on one struct
so far.